+++ frontend/package.json
@@ -15,6 +15,8 @@
         "react": "16.8.6",
         "react-bootstrap": "^1.0.0-beta.12",
         "react-dom": "16.8.6",
+        "react-router": "^5.1.2",
+        "react-router-dom": "^5.1.2",
         "react-scripts": "2.1.1",
         "typescript": "3.5.2"
     },
