+++ frontend/src/components/portfolios/portfolio/portfolio.tsx
@@ -14,6 +14,10 @@ export interface PortfolioProps {
   onPortfolioChange: (portfolio: PortfolioModel) => void;
 }
 
+// const useTransactionFormSubmit = (value: Transaction) => {
+//   return useEff
+// }
+
 export function Portfolio(props: PortfolioProps) {
 
   const { portfolio } = props;
